+++ src/structs.rs
@@ -21,9 +21,9 @@ use tabled::{settings::Style, Table, Tabled};
 use crate::{
     OPT_AVAILABILITY_INTERVAL, OPT_AVAILABILITY_WINDOW, OPT_COLUMNS, OPT_EXCLUDE_CHANNEL_STATES,
     OPT_FLOW_STYLE, OPT_FORWARDS, OPT_FORWARDS_ALIAS, OPT_FORWARDS_FILTER_AMT,
-    OPT_FORWARDS_FILTER_FEE, OPT_INVOICES, OPT_INVOICES_FILTER_AMT, OPT_JSON, OPT_LOCALE,
-    OPT_MAX_ALIAS_LENGTH, OPT_MAX_DESC_LENGTH, OPT_PAYS, OPT_PAYS_COLUMNS, OPT_REFRESH_ALIAS,
-    OPT_SORT_BY, OPT_STYLE, OPT_UTF8,
+    OPT_FORWARDS_FILTER_FEE, OPT_INVOICES, OPT_INVOICES_COLUMNS, OPT_INVOICES_FILTER_AMT, OPT_JSON,
+    OPT_LOCALE, OPT_MAX_ALIAS_LENGTH, OPT_MAX_DESC_LENGTH, OPT_PAYS, OPT_PAYS_COLUMNS,
+    OPT_REFRESH_ALIAS, OPT_SORT_BY, OPT_STYLE, OPT_UTF8,
 };
 
 pub const NO_ALIAS_SET: &str = "NO_ALIAS_SET";
@@ -43,6 +43,8 @@ pub struct Config {
     pub pays_columns: DynamicConfigOption<Vec<String>>,
     pub max_desc_length: DynamicConfigOption<i64>,
     pub invoices: DynamicConfigOption<u64>,
+    pub invoices_columns: DynamicConfigOption<Vec<String>>,
+    pub max_label_length: DynamicConfigOption<i64>,
     pub invoices_filter_amt_msat: DynamicConfigOption<i64>,
     pub locale: DynamicConfigOption<Locale>,
     pub refresh_alias: DynamicConfigOption<u64>,
@@ -113,6 +115,20 @@ impl Config {
             invoices: DynamicConfigOption {
                 name: OPT_INVOICES,
                 value: 0,
+            },
+            invoices_columns: DynamicConfigOption {
+                name: OPT_INVOICES_COLUMNS,
+                value: {
+                    Invoices::FIELD_NAMES_AS_ARRAY
+                        .into_iter()
+                        .filter(|t| t != &"description" && t != &"preimage")
+                        .map(|s| s.to_string())
+                        .collect::<Vec<String>>()
+                },
+            },
+            max_label_length: DynamicConfigOption {
+                name: OPT_MAX_DESC_LENGTH,
+                value: 30,
             },
             invoices_filter_amt_msat: DynamicConfigOption {
                 name: OPT_INVOICES_FILTER_AMT,
@@ -313,16 +329,24 @@ pub struct Pays {
     pub preimage: String,
 }
 
-#[derive(Debug, Tabled, Serialize)]
+#[derive(Debug, Tabled, FieldNamesAsArray, Serialize)]
 pub struct Invoices {
     #[tabled(skip)]
     pub paid_at: u64,
     #[tabled(rename = "paid_at")]
     #[serde(skip_serializing)]
+    #[field_names_as_array(skip)]
     pub paid_at_str: String,
     pub label: String,
-    #[tabled(rename = "sats_received")]
+    #[serde(skip_serializing_if = "String::is_empty")]
+    pub description: String,
+    #[tabled(skip)]
+    #[field_names_as_array(skip)]
     pub msats_received: u64,
+    #[serde(skip_serializing)]
+    pub sats_received: u64,
+    pub payment_hash: String,
+    pub preimage: String,
 }
 
 #[derive(Debug, Clone, Default)]